+++ src/test/java/net/byteseek/matcher/bytes/SetBinarySearchMatcherTest.java
@@ -169,11 +169,11 @@ public class SetBinarySearchMatcherTest {
         assertFalse(matcher.matches(reader, -1L));
         assertFalse(matcher.matches(reader, 256L));
         for (Byte byteShouldMatch : bytesMatched) {
-            long bytePosition = byteShouldMatch.byteValue() & 0xff;
+            long bytePosition = byteShouldMatch & 0xff;
             assertEquals(String.format("%s: Byte %02x should match:", description, byteShouldMatch), true, matcher.matches(reader, bytePosition));
         }
         for (Byte byteShouldNotMatch : bytesNotMatched) {
-            long bytePosition = byteShouldNotMatch.byteValue() & 0xff;
+            long bytePosition = byteShouldNotMatch & 0xff;
             assertEquals(String.format("%s: Byte %02x should not match:", description, byteShouldNotMatch), false, matcher.matches(reader, bytePosition));
         }
 
@@ -181,11 +181,11 @@ public class SetBinarySearchMatcherTest {
         assertFalse(matcher.matches(BYTE_VALUES, -1));
         assertFalse(matcher.matches(BYTE_VALUES, 256));
         for (Byte byteShouldMatch : bytesMatched) {
-            int bytePosition = byteShouldMatch.byteValue() & 0xff;
+            int bytePosition = byteShouldMatch & 0xff;
             assertEquals(String.format("%s: Byte %02x should match:", description, byteShouldMatch), true, matcher.matches(BYTE_VALUES, bytePosition));
         }
         for (Byte byteShouldNotMatch : bytesNotMatched) {
-            int bytePosition = byteShouldNotMatch.byteValue() & 0xff;
+            int bytePosition = byteShouldNotMatch & 0xff;
             assertEquals(String.format("%s: Byte %02x should not match:", description, byteShouldNotMatch), false, matcher.matches(BYTE_VALUES, bytePosition));
         }
 
@@ -201,11 +201,11 @@ public class SetBinarySearchMatcherTest {
         } catch (ArrayIndexOutOfBoundsException expectedIgnore) {}
 
         for (Byte byteShouldMatch : bytesMatched) {
-            int bytePosition = byteShouldMatch.byteValue() & 0xff;
+            int bytePosition = byteShouldMatch & 0xff;
             assertEquals(String.format("%s: Byte %02x should match:", description, byteShouldMatch), true, matcher.matchesNoBoundsCheck(BYTE_VALUES, bytePosition));
         }
         for (Byte byteShouldNotMatch : bytesNotMatched) {
-            int bytePosition = byteShouldNotMatch.byteValue() & 0xff;
+            int bytePosition = byteShouldNotMatch & 0xff;
             assertEquals(String.format("%s: Byte %02x should not match:", description, byteShouldNotMatch), false, matcher.matchesNoBoundsCheck(BYTE_VALUES, bytePosition));
         }
